Update output to match our specs
python 2.x matched these correctly without the casts but a version
bump to python 3.x (identified in 3.5.3) changes the output, this
change casts to int when we expect an integer output

+++ nova/virt/lxd/driver.py
@@ -897,10 +897,10 @@ class LXDDriver(driver.ComputeDriver):
 
         data = {
             'vcpus': vcpus,
-            'memory_mb': local_memory_info['total'] / units.Mi,
-            'memory_mb_used': local_memory_info['used'] / units.Mi,
-            'local_gb': local_disk_info['total'] / units.Gi,
-            'local_gb_used': local_disk_info['used'] / units.Gi,
+            'memory_mb': int(local_memory_info['total'] / units.Mi),
+            'memory_mb_used': int(local_memory_info['used'] / units.Mi),
+            'local_gb': int(local_disk_info['total'] / units.Gi),
+            'local_gb_used': int(local_disk_info['used'] / units.Gi),
             'vcpus_used': 0,
             'hypervisor_type': 'lxd',
             'hypervisor_version': '011',
